Showcontroller edit

+++ src/main/java/com/kinoxp/repositories/ShowRepository.java
@@ -3,5 +3,5 @@ package com.kinoxp.repositories;
 import com.kinoxp.entities.Show;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShowRepository extends JpaRepository<Show, Integer> {
+public interface ShowRepository extends JpaRepository<Show, Long> {
 }
